+++ test/down_test.rb
@@ -100,19 +100,22 @@ describe Down do
     end
 
     it "follows redirects" do
-      stub_request(:get, "http://example.com").to_return(status: 301, headers: {'Location' => 'http://example1.com'})
-      stub_request(:get, "http://example1.com").to_return(status: 301, headers: {'Location' => 'http://example2.com'})
-
-      stub_request(:get, "http://example2.com").to_return(body: "a" * 5)
+      # Double redirect
+      stub_request(:get, "http://example.com").to_return(status: 301, headers: {'Location' => '/foo'}) # relative
+      stub_request(:get, "http://example.com/foo").to_return(status: 301, headers: {'Location' => 'https://example.com/bar'}) # HTTP => HTTPS
+      stub_request(:get, "https://example.com/bar").to_return(body: "file", headers: {"Content-Type" => "text/plain"})
       tempfile = Down.download("http://example.com")
-      assert_equal "aaaaa", tempfile.read
+      assert_equal "file", tempfile.read
+      assert_equal "bar", tempfile.original_filename
+      assert_equal "text/plain", tempfile.content_type
 
-      stub_request(:get, "http://example2.com").to_return(status: 301, headers: {'Location' => 'http://example3.com'})
-      assert_raises(Down::NotFound) { Down.download("http://example.com") }
-
-      stub_request(:get, "http://example3.com").to_return(body: "a" * 5)
+      stub_request(:get, "http://example.com").to_return(status: 301, headers: {'Location' => 'http://example.com/foo'})
+      stub_request(:get, "http://example.com/foo").to_return(status: 301, headers: {'Location' => 'http://example.com/bar'})
+      stub_request(:get, "http://example.com/bar").to_return(status: 301, headers: {'Location' => 'http://example.com/baz'})
+      stub_request(:get, "http://example.com/baz").to_return(body: "file")
+      assert_raises(Down::Error) { Down.download("http://example.com") }
       tempfile = Down.download("http://example.com", max_redirects: 3)
-      assert_equal "aaaaa", tempfile.read
+      assert_equal "file", tempfile.read
     end
 
     it "preserves extension" do
@@ -137,6 +140,13 @@ describe Down do
 
       tempfile = Down.download("http://user:password@example.com/image.jpg")
       assert_equal "aaaaa", tempfile.read
+    end
+
+    it "accepts a proxy" do
+      stub_request(:get, "http://example.com/image.jpg")
+
+      Down.download("http://example.com/image.jpg", proxy: "http://proxy.org")
+      Down.download("http://example.com/image.jpg", proxy: "http://user:password@proxy.org")
     end
 
     it "forwards options to open-uri" do
@@ -238,6 +248,33 @@ describe Down do
         .to_return(body: "abc")
 
       Down.open("http://example.com/image.jpg", {"Header" => "value"})
+    end
+
+    it "accepts a proxy" do
+      stub_request(:get, "http://example.com/image.jpg")
+
+      Down.open("http://example.com/image.jpg", proxy: "http://proxy.org")
+      Down.open("http://example.com/image.jpg", proxy: "http://user:password@proxy.org")
+    end
+
+    it "saves the response status and headers" do
+      stub_request(:get, "http://example.com/image.jpg")
+        .to_return(body: "abc", headers: {"My-Header" => "Value"})
+
+      io = Down.open("http://example.com/image.jpg")
+
+      assert_equal 200, io.data[:status]
+      assert_equal Hash["My-Header" => "Value"], io.data[:headers]
+    end
+
+    it "raises an error on 4xx and 5xx response" do
+      stub_request(:get, "http://example.com/image.jpg").to_return(status: 404, body: "File not found")
+      error = assert_raises(Down::Error) { Down.open("http://example.com/image.jpg") }
+      assert_match "request to http://example.com/image.jpg returned status 404 and body:\nFile not found", error.message
+
+      stub_request(:get, "http://example.com/image.jpg").to_return(status: 500, body: "There has been an error")
+      error = assert_raises(Down::Error) { Down.open("http://example.com/image.jpg") }
+      assert_equal "request to http://example.com/image.jpg returned status 500 and body:\nThere has been an error", error.message
     end
   end